# CoinCollectionFrontend/src/api/models/HistoryEntryByCoinDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HistoryEntryByCoinDto = {
    dateTime?: string;
    entryValue?: number;
};