# CoinCollectionFrontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CoinCollectionBackend.Web",
    "version": "1.0"
  },
  "paths": {
    "/Auth": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/Coins/all": {
      "get": {
        "tags": [
          "Coins"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Coins/by-id/{coinId}": {
      "get": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              }
            }
          }
        }
      }
    },
    "/Coins/by-currency/{currencyId}": {
      "get": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "currencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Coins/by-currencies": {
      "get": {
        "tags": [
          "Coins"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinGroupByCurrencyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Coins": {
      "post": {
        "tags": [
          "Coins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Coins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CoinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoinDto"
                }
              }
            }
          }
        }
      }
    },
    "/Coins/by-id/{coinId}/front": {
      "get": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Coins/by-id/{coinId}/back": {
      "get": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Currencies/all": {
      "get": {
        "tags": [
          "Currencies"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Currencies/by-id/{currencyId}": {
      "get": {
        "tags": [
          "Currencies"
        ],
        "parameters": [
          {
            "name": "currencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyDto"
                }
              }
            }
          }
        }
      }
    },
    "/History/by-id/{coinId}": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntryByCoinDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/History/by-id/{coinId}/updates": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoinDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CoinGroupByCurrencyDto": {
        "type": "object",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoinGroupByCurrencyItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CoinGroupByCurrencyItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CurrencyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HistoryEntryByCoinDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "entryValue": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}